convert scripts to ts
